Replaced 1 by 1L, etc, in various places.
   NAMESPACE (auto-generated)
   	Format of function declarations has been neatened up.

   fardist.R
   fasp.R
   fgk3.R
   fiksel.R
   fitted.ppm.R
   flipxy.R
   formulae.R
   fourierbasis.R
   fryplot.R
   fv.R
   geyer.R
   hackglmm.R
   hermite.R
   hexagons.R
   hierhard.R
   hierpair.family.R
   hierstrhard.R
   hybrid.family.R
   hybrid.R
   hyperframe.R
   hypersub.R
   idw.R
   images.R
   indices.R
   infline.R
   interp.im.R
   iplotlayered.R
   iplot.R
   ippm.R
   is.cadlag.R
   is.subset.owin.R
   kernel2d.R
   kppm.R
   laslett.R
   layered.R
   lennard.R
   leverage.R
   linalg.R
   lineardisc.R
   linearKmulti.R
   linearmrkcon.R
   linearpcfmulti.R
   linequad.R
   linfun.R
   linim.R
   linnet.R
   linnetsurgery.R
   lintess.R
   lixellate.R
   logistic.R
   lohboot.R
   lppm.R
   lpp.R
   lurking.R
	Replaced 1 by 1L, etc

   pixellate.R
   polygood.R
   primefactors.R
   pspcross.R
   rmh.default.R
   rmhsnoop.R
   strauss.R
   tapplysum.R
   unique.ppp.R
   window.R
   wingeom.R
	Tweaks related to integers

   DESCRIPTION
   NEWS
	Updated as version 1.48-0.038

+++ R/areainter.R
@@ -200,7 +200,7 @@ areadelta2 <- local({
                 yother = as.double(yy[K]),
                 radius = as.double(r),
                 epsilon = as.double(eps),
-                pixcount = as.integer(integer(1)))
+                pixcount = as.integer(integer(1L)))
         result[i,j] <- result[j,i] <- z$pixcount
       }
       # normalise
@@ -345,7 +345,7 @@ areadelta2 <- local({
             yother = as.double(yy[K]),
             radius = as.double(r),
             epsilon = as.double(eps),
-            pixcount = as.integer(integer(1)))
+            pixcount = as.integer(integer(1L)))
       result[i,j] <- z$pixcount
     }
     # normalise
